move package reader writer to local
