+++ src/main/kotlin/data/source/local/csv_reader_writer/audits/AuditCSVReaderWriter.kt
@@ -1,6 +1,6 @@
-package org.example.data.utils.csv_reader_writer.mate_task_assignment
+package org.example.data.source.local.csv_reader_writer.audits
 
-import data.dto.MateTaskAssignmentDto
+import data.dto.AuditDto
 import org.jetbrains.kotlinx.dataframe.DataFrame
 import org.jetbrains.kotlinx.dataframe.api.cast
 import org.jetbrains.kotlinx.dataframe.api.concat
@@ -11,11 +11,10 @@ import org.jetbrains.kotlinx.dataframe.io.writeCSV
 import java.io.File
 
 
-class MateTaskAssignmentCSVReaderWriter(private val filePath: String) : IMateTaskAssignmentCSVReaderWriter {
-
+class AuditCSVReaderWriter(private val filePath: String) : IAuditCSVReaderWriter {
     private fun resolveFile(): File = File(filePath)
 
-    override suspend fun read(): List<MateTaskAssignmentDto> {
+    override suspend fun read(): List<AuditDto> {
         val file = resolveFile()
         if (!file.exists()) {
             file.createNewFile()
@@ -27,24 +26,24 @@ class MateTaskAssignmentCSVReaderWriter(private val filePath: String) : IMateTas
         }
 
         return DataFrame.readCSV(file)
-            .cast<MateTaskAssignmentDto>()
+            .cast<AuditDto>()
             .toList()
     }
 
-    override suspend fun overWrite(mateTasks: List<MateTaskAssignmentDto>): Boolean {
-        mateTasks.toDataFrame().writeCSV(resolveFile())
+    override suspend fun overWrite(audits: List<AuditDto>): Boolean {
+        audits.toDataFrame().writeCSV(resolveFile())
         return true
     }
 
-    override suspend fun append(mateTasks: List<MateTaskAssignmentDto>): Boolean {
+    override suspend fun append(audits: List<AuditDto>): Boolean {
         val file = resolveFile()
         val existing = if (file.exists() && file.length() > 0) {
             DataFrame.readCSV(file).cast()
         } else {
-            emptyList<MateTaskAssignmentDto>().toDataFrame()
+            emptyList<AuditDto>().toDataFrame()
         }
 
-        val newData = mateTasks.toDataFrame()
+        val newData = audits.toDataFrame()
         (existing.concat(newData)).writeCSV(file)
         return true
     }
