+++ src/main/kotlin/data/source/local/csv_reader_writer/audits/IAuditCSVReaderWriter.kt
@@ -1,4 +1,4 @@
-package org.example.data.utils.csv_reader_writer.audits
+package org.example.data.source.local.csv_reader_writer.audits
 
 import data.dto.AuditDto
