+++ src/main/kotlin/data/source/local/csv_reader_writer/project/IProjectCSVReaderWriter.kt
@@ -1,4 +1,4 @@
-package org.example.data.utils.csv_reader_writer.project
+package org.example.data.source.local.csv_reader_writer.project
 
 import data.dto.ProjectDto
