+++ src/main/kotlin/data/source/local/csv_reader_writer/state/StateCSVReaderWriter.kt
@@ -1,6 +1,6 @@
-package org.example.data.utils.csv_reader_writer.user_assigned_to_project
+package org.example.data.source.local.csv_reader_writer.state
 
-import data.dto.UserAssignedToProjectDto
+import data.dto.ProjectStateDto
 import org.jetbrains.kotlinx.dataframe.DataFrame
 import org.jetbrains.kotlinx.dataframe.api.cast
 import org.jetbrains.kotlinx.dataframe.api.concat
@@ -10,12 +10,12 @@ import org.jetbrains.kotlinx.dataframe.io.readCSV
 import org.jetbrains.kotlinx.dataframe.io.writeCSV
 import java.io.File
 
-class UserAssignedToProjectCSVReaderWriter(private val filePath: String) : IUserAssignedToProjectCSVReaderWriter {
+class StateCSVReaderWriter(private val filePath: String) :
+    org.example.data.source.local.csv_reader_writer.state.IStateCSVReaderWriter {
     private fun resolveFile(): File = File(filePath)
 
-    override suspend fun read(): List<UserAssignedToProjectDto> {
+    override suspend fun read(): List<ProjectStateDto> {
         val file = resolveFile()
-
         if (!file.exists()) {
             file.createNewFile()
         }
@@ -24,25 +24,26 @@ class UserAssignedToProjectCSVReaderWriter(private val filePath: String) : IUser
             return emptyList()
 
         val users = DataFrame.readCSV(file)
-            .cast<UserAssignedToProjectDto>()
+            .cast<ProjectStateDto>()
             .toList()
         return users
+
     }
 
-    override suspend fun overWrite(users: List<UserAssignedToProjectDto>): Boolean {
-        users.toDataFrame().writeCSV(resolveFile())
+    override suspend fun overWrite(state: List<ProjectStateDto>): Boolean {
+        state.toDataFrame().writeCSV(resolveFile())
         return true
     }
 
-    override suspend fun append(users: List<UserAssignedToProjectDto>): Boolean {
+    override suspend fun append(state: List<ProjectStateDto>): Boolean {
         resolveFile().also { file ->
             val existing = if (file.exists() && file.length() > 0) {
                 DataFrame.readCSV(file).cast()
-            } else emptyList<UserAssignedToProjectDto>().toDataFrame()
+            } else emptyList<ProjectStateDto>().toDataFrame()
 
-            val newData = users.toDataFrame()
+            val newData = state.toDataFrame()
             (existing.concat(newData)).writeCSV(file)
+            return true
         }
-        return true
     }
 }
