+++ src/main/kotlin/data/source/local/csv_reader_writer/user_assigned_to_project/UserAssignedToProjectCSVReaderWriter.kt
@@ -1,6 +1,6 @@
-package org.example.data.utils.csv_reader_writer.state
+package org.example.data.source.local.csv_reader_writer.user_assigned_to_project
 
-import data.dto.ProjectStateDto
+import data.dto.UserAssignedToProjectDto
 import org.jetbrains.kotlinx.dataframe.DataFrame
 import org.jetbrains.kotlinx.dataframe.api.cast
 import org.jetbrains.kotlinx.dataframe.api.concat
@@ -10,12 +10,12 @@ import org.jetbrains.kotlinx.dataframe.io.readCSV
 import org.jetbrains.kotlinx.dataframe.io.writeCSV
 import java.io.File
 
-class StateCSVReaderWriter(private val filePath: String) :
-    org.example.data.utils.csv_reader_writer.state.IStateCSVReaderWriter {
+class UserAssignedToProjectCSVReaderWriter(private val filePath: String) : IUserAssignedToProjectCSVReaderWriter {
     private fun resolveFile(): File = File(filePath)
 
-    override suspend fun read(): List<ProjectStateDto> {
+    override suspend fun read(): List<UserAssignedToProjectDto> {
         val file = resolveFile()
+
         if (!file.exists()) {
             file.createNewFile()
         }
@@ -24,26 +24,25 @@ class StateCSVReaderWriter(private val filePath: String) :
             return emptyList()
 
         val users = DataFrame.readCSV(file)
-            .cast<ProjectStateDto>()
+            .cast<UserAssignedToProjectDto>()
             .toList()
         return users
-
     }
 
-    override suspend fun overWrite(state: List<ProjectStateDto>): Boolean {
-        state.toDataFrame().writeCSV(resolveFile())
+    override suspend fun overWrite(users: List<UserAssignedToProjectDto>): Boolean {
+        users.toDataFrame().writeCSV(resolveFile())
         return true
     }
 
-    override suspend fun append(state: List<ProjectStateDto>): Boolean {
+    override suspend fun append(users: List<UserAssignedToProjectDto>): Boolean {
         resolveFile().also { file ->
             val existing = if (file.exists() && file.length() > 0) {
                 DataFrame.readCSV(file).cast()
-            } else emptyList<ProjectStateDto>().toDataFrame()
+            } else emptyList<UserAssignedToProjectDto>().toDataFrame()
 
-            val newData = state.toDataFrame()
+            val newData = users.toDataFrame()
             (existing.concat(newData)).writeCSV(file)
-            return true
         }
+        return true
     }
 }
